_readelf_fd: try MAP_SHARED to see if it performs better than MAP_PRIVATE
Bug 749624 reports that scanelf has a slow performance problem
when processing large numbers of non-elf files which are backed
by some form of slow network storage. It may be that the
performance problem it due to mmap usage in the _readelf_fd,
therefore try MAP_SHARED to see if it performs better than
MAP_PRIVATE for this slow network storage scenario.

Bug: https://bugs.gentoo.org/749624